fix(memory_collector): adjust params type of __init__
adjust type of param `monitoring_interval` in function __ini__ from int
into float.

+++ src/mymonitor/memory_collectors/base.py
@@ -45,7 +45,7 @@ class AbstractMemoryCollector(ABC):
     matching a given pattern.
     """
 
-    def __init__(self, process_pattern: str, monitoring_interval: int, **kwargs):
+    def __init__(self, process_pattern: str, monitoring_interval: float, **kwargs):
         """
         Initializes the AbstractMemoryCollector.
